# ecommerce-platform/services/auth-service/src/main/java/com/ecommerce/authservice/api/dto/AuthResponse.java
package com.ecommerce.authservice.api.dto;

public record AuthResponse(String accessToken, String refreshToken, String tokenType) {

    public static AuthResponse of(String accessToken, String refreshToken) {
        return new AuthResponse(accessToken, refreshToken, "Bearer");
    }
}