# ecommerce-platform/services/auth-service/src/main/java/com/ecommerce/authservice/api/dto/RegisterRequest.java
package com.ecommerce.authservice.api.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8, max = 100) String password) {
}